update client utils folder structure
