+++ client/src/utils/api/index.js
@@ -1,0 +1,2 @@
+export { createModel, updateModel, deleteModel } from './modelCRUD'
+export { loginWithPassword, loginWithToken } from './userAuth'
