+++ client/src/utils/api/modelCRUD/index.js
@@ -1,0 +1,3 @@
+export { createModel } from './create'
+export { updateModel } from './update'
+export { deleteModel } from './delete'
