+++ client/src/utils/api/userAuth/index.js
@@ -1,0 +1,2 @@
+export { loginWithPassword } from './loginWithPassword'
+export { loginWithToken } from './loginWithToken'
