+++ client/src/utils/helpers/index.js
@@ -1,0 +1,7 @@
+export { logoutTutor } from './logout'
+export {
+    validateEmail,
+    validatePassword,
+    validateFormInputs,
+    validateSelect
+} from './forms'
