+++ client/src/utils/helpers/logout.js
