+++ client/src/utils/index.js
@@ -1,14 +1,16 @@
-import { loginWithToken } from "./loginWithToken";
-import { loginWithPassword } from "./loginWithPassword";
-import { logoutTutor } from "./logout";
-import { createModel, deleteModel, updateModel } from "./models";
-
 export {
     loginWithToken,
     loginWithPassword,
-    logoutTutor,
 
     createModel,
-    updateModel,
-    deleteModel
-}
+    deleteModel,
+    updateModel
+} from "./api";
+
+export {
+    logoutTutor,
+    validateEmail,
+    validatePassword,
+    validateFormInputs,
+    validateSelect
+} from "./helpers";
